asdf

+++ map/map_exception.c
@@ -6,18 +6,18 @@
 /*   By: jihong <jihong@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/01 23:24:27 by jihong            #+#    #+#             */
-/*   Updated: 2022/07/02 00:16:36 by jihong           ###   ########.fr       */
+/*   Updated: 2022/07/02 15:59:44 by jihong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../so_long.h"
 
-void check_wall_exception(t_game_attribute *g)
+void	check_wall_exception(t_game_attribute *g)
 {
 	int	i;
 
 	i = 0;
-	while (i < ft_strlen(g->map_str))
+	while (i < (int)ft_strlen(g->map_str))
 	{
 		if (i < g->width)
 		{
@@ -29,7 +29,7 @@ void check_wall_exception(t_game_attribute *g)
 			if (g->map_str[i] != '1')
 				error_massage("check wall\n");
 		}
-		else if (i > ft_strlen(g->map_str) - g->width)
+		else if (i > (int)ft_strlen(g->map_str) - g->width)
 		{
 			if (g->map_str[i] != '1')
 				error_massage("check wall\n");
@@ -38,16 +38,16 @@ void check_wall_exception(t_game_attribute *g)
 	}
 }
 
-void check_map_rectangle(t_game_attribute *g)
+void	check_map_rectangle(t_game_attribute *g)
 {
-	if (g->width * g->height != ft_strlen(g->map_str))
+	if (g->width * g->height != (int)ft_strlen(g->map_str))
 		error_massage("check map!!!!");
 }
 
-void check_attribute_exception(t_game_attribute *g)
+void	check_attribute_exception(t_game_attribute *g)
 {
 	int	palyer_num;
-	int door_num;
+	int	door_num;
 	int	i;
 
 	i = 0;
@@ -55,7 +55,7 @@ void check_attribute_exception(t_game_attribute *g)
 	door_num = 0;
 	g->col = 0;
 	g->col_have = 0;
-	while (i ++ < ft_strlen(g->map_str))
+	while (i ++ < (int)ft_strlen(g->map_str))
 	{
 		if (g->map_str[i] == 'E')
 			door_num ++;
@@ -72,7 +72,7 @@ void check_attribute_exception(t_game_attribute *g)
 		error_massage("check collection");
 }
 
-void check_all_exception(t_game_attribute *game)
+void	check_all_exception(t_game_attribute *game)
 {
 	check_wall_exception(game);
 	check_map_rectangle(game);
